requirements.txt
requests
colorama
Flask
cloudinary
python-dotenv
